+++ src/main/java/com/management/tab/application/auth/LoginService.java
@@ -1,7 +1,7 @@
 package com.management.tab.application.auth;
 
 import com.management.tab.application.auth.dto.LoggedInUserDto;
-import com.management.tab.config.auth.security.repository.UserSocialRepository;
+import com.management.tab.domain.repository.UserSocialRepository;
 import com.management.tab.domain.user.User;
 import com.management.tab.domain.user.vo.RegistrationId;
 import com.management.tab.domain.user.vo.Social;
