+++ src/main/java/com/management/tab/persistence/JdbcUserRepository.java
@@ -1,6 +1,6 @@
 package com.management.tab.persistence;
 
-import com.management.tab.config.auth.security.repository.UserSocialRepository;
+import com.management.tab.domain.repository.UserSocialRepository;
 import com.management.tab.domain.repository.UserRepository;
 import com.management.tab.domain.user.User;
 import com.management.tab.domain.user.vo.Social;
